Update README; Include `log` directory
Update README to reflect that JDK 8 and sbt are both required and
sbt will be used to test and package the app into a JAR file.

The `log` directory was added to the empty (though all *.log files
are ignored) so that a user wouldn't have to manually add the `log`
directory after cloning the repo.