Add upload package to upload directly to S3 bucket
Add upload package to make it easier to upload zipped DICOM
directories directly to a target S3 bucket and key prefix as
defined in `src/main/resources/config/upload.conf` (or
`src/main/resources/config/upload_test.conf` when testing on
non-HIPAA compliant S3 bucket).

Using the AWS SDK for Java v1 rquired a few more JARs to be added
to `lib/`, which are now included in this repo.

Most of the README.md was updated to reflect the additional upload
package, but this still needs a thorough review and probably some
light editing.